Fix Linux C4244 warnings: '<usage>': conversion from '<biggertype>' to '<smallertype>', possible loss of data #review-11113622 Devin.Doucette #rnx #rb trivial
[CL 11113621 by Matt Peters in Dev-Core branch]

+++ Engine/Source/Runtime/Core/Private/String/HexToBytes.cpp
@@ -25,7 +25,7 @@ namespace String
 			const uint8 HiNibble = TCharToNibble(*HexPos++) << 4;
 			*OutPos++ = HiNibble | TCharToNibble(*HexPos++);
 		}
-		return OutPos - OutBytes;
+		return static_cast<int32>(OutPos - OutBytes);
 	}
 
 	int32 HexToBytes(const FStringView& Hex, uint8* OutBytes)
